\name{gene.gene.statistic}
\alias{gene.gene.statistic}
\title{ Compute gene-gene statistics }
\description{
 Computes gene gene statistics.
}
\usage{
gene.gene.statistic(g1, g2, numPapers, PaperLen)
}
\arguments{
  \item{g1}{The Entrez Gene identifier for one of the genes. }
  \item{g2}{The Entrez Gene identifier for the other gene. }
  \item{numPapers}{The number of papers being used. }
  \item{PaperLen}{ A vector of length \code{numPapers} with the 
   number of citations for each paper.  }
}
\details{
To be filled in later.
}
\value{
To be filled in later.
}
\references{ The Wiley Book chapter. }
\author{ B. Ding}
\seealso{ \code{\link{twowayTable}} }
\examples{
}
\keyword{ manip }
