\name{gene.geneslist.statistic}
\alias{gene.geneslist.statistic}
\title{A function to compute association measures between a gene of interest
and a list of genes. }
\description{
 Whether or not a gene has an association with another gene, or a set of 
genes is measured using co-citation in PubMed as a basis for measuring
that association.
}
\usage{
gene.geneslist.statistic(gene, geneslist, paperLens)
}
\arguments{
  \item{gene}{The Entrez Gene identifier for the gene of interest. }
  \item{geneslist}{A vector of Entrez Gene identifiers for the set of genes. }
  \item{paperLens}{A vector containing the number of genes cited by each 
   paper.}
}
\details{
To be filled in later.
}
\value{
To be filled in later.
}
\references{ The Springer Book. }
\author{B. Ding }

\seealso{ \code{\link{twowayTable}}}
\examples{
}
\keyword{ manip }
